all in one

+++ seti/transforms.py
@@ -100,7 +100,7 @@ class GaussianNoise(torch.nn.Module):
             PIL Image or Tensor: Randomly flipped image.
         """
         if torch.rand(1) < self.p:
-            img = torch.tensor(random_noise(np.array(img), mode='gaussian', mean=0, var=0.001, clip=True))
+            img = torch.tensor(random_noise(np.array(img), mode='gaussian', mean=0, var=0.001, clip=True), dtype=torch.Float)
             return img
         return img
